+++ accounts.py
@@ -2,37 +2,23 @@ from pymongo import MongoClient
 from config import *
 import utils
 import mongo_client
-#[phone,password]
-id_list=[[b'\xa0a\xa6^\xc7\x0cF|\xe9\xf2\xa1\xd8fN\x95s\xbd\xfcA\xebl0\xd2\xf9\xa9\xdb\xe5\xc0\xfc\x9b\tf\xcd\x1c|q\x17\xb9L\x03\xa6\x91?&\x1d\x1a\x99\x00\xb7\xe1\xf1\xe5J_v/n\xdb\x8d\xc3\xd3TL\x89\x8a\xae \xdf\x97\x91Kqy\x84U\x03\xe8\xd9A\xec\xe6\xa8\xf7c\xbf\xc8\xe1W\xb7\xb8\x9e\xb3\xc2\xfe\x93\xe0\x1c\x08\x86\xfb\x9b\xbe\x08\x1e\xb8\x11\x00\x15\x14dkE\xe6^\xf1\xa6$\xe3\x0b\xd5;\x97G&K\xc1\x85\x85', b'c\xe5\xde\r\x88\x05\xdc\xb9\xac\xf6\xd4d\x02\xe7\x00(y5RF$\xc4;\xef]_\xbcR`\x8e\xf8\xc3\'\xd1l\xe7\xa0\x90\xd3\xae\x8c\xe2\x146"}\xa6\xfe\xa1\xe6\xf8[\xae\x9c\x99\x1a\x9c\xd1N.\x80Uo7\xa8?\xba\xaa\xab\x80g\x9c\xfb!\xb3\x12\x80\xa3\xd8\xcc\xb9\x8dU+\xcb\xcc\xe8_\x07\x99\xcb;Xk\x93\x9f(\xab\x89\x9c\x1cH\xe6\xd7\x8bh!z\x1b\x8b~\xe1j\xfc\xa6\xb0\xfbno;\xea\x1d\x83,t\x8bo\xaf'],
-[b'\x15;b\xa1\x98[\xa8\xeeQ\xf14e\x9a\xd4\xf9\t\xf7\x87\xcf\xe9\x0c\xf40)\xed5&\x13Km\xab\xd9\xb3\xdc\xe1_\xd2\xa5\xf6\t\x96\xa7n\x94;\x02\xdf\xcc\x85\xb0\xba\xcf\x8f\x04\x81\xd0\xfb\x99\x19\x11hP\xc8\xb2\x07\xe7\xa23\xd3\xd5r\x88\xfe\x97\xc7\x0bF\xb81Kz(\xc2_\x8b\x81\x15\xba\x8aP\x05\x1f\x00Q\xf6Y\x8b\xdaA\xb7\xb2\x97t\xf0\xc7\xfd\x94\xe7\x1c\x13\xdc\xe6\x7f<\xc7v\xa0W\xe5d\xf2$u\x00J\xa7\xf0\x90', b'\\\xc8\xb7\xff\x818\xa2\x8b0\x01\xa99\xa2\xb4M\xfcyq\xd2*0\xc9\xbe\xb2\x8ak\xf4\xa7\xd1YT,\x04\xf8X\x8c\xcaF\xb8\xc7@\xd1\x16G\x8a^\xff\xfa\xb4\xd8\xc8\xcb\x81;\x14\xaa^\xc4\xeb\xda\xfe\xce\xf9I\xdd\x8c\xd3\xd0\xafX\xacRH}Y\x86\x1fw\x15D\xff\x13\xe5\r\x81\x12\xbc\xd0\xd9\x14\xb0XJ\x90vjj\xf6\x0e\xa0G]%\x12\xb9\x9a{1\x917\x98\xfb\x19\x15\xf6\x08\xee\xf5\x1a\xe95\x05\x9c\xef\xb0\xd9\x92|'],
-[b'o\xcd\x15\n\xccIW:?\xa1y\x8d\xf0\xe1Yz\x92\x85\xfb\xec\xcfQ\x94\xd4\xd3\xd2\xeei\xe0Tc\xe3R^$\xcdv3\x89\xcf\xbf\xdb\x8a(\x1c\xea:\x0e\x06\xbe\x8aV\xa2\xac\xf3?\x98\xc9+\x96\xaf\xd4Qg7\xe7\x16\xb1"\x84\xf2\t\x10\xce&\xbf\xdfKk_7\x90\x80\xbe83\x9d\xfa"\x96\xa0x\x10C\xf1\xc3\xab\xe8\xe4l~A\xa0\xb5:@\xd9Q\x81/&m7\xa6\xba\xd1\x8a?\x07\x8c\xbb`p\x828\tp\xae', b'\x13\xabT\x0f7&\x91\xde\xca\xc8.\xeb\x0c\xbb\xf1<\xb0\xca;g\xc5\x02g\xa5\x0e\x80\xa2\xcd\xb0@\x9d\t\x92\x0c\x00\xb8\xb0\xaf |F\x99V.{\x8b\x90H!\xd3\xe1r}^-\x9e\x93\xa7\xea\x044\xdc\xb9/H\x98X\xbb\\r\xa7R\xd4y\xf3\x9b\x94\x9b\xb9\xde]\xa1\xb6o\x91\xb4\x1bq\xb6\xf6WT\xc8\xf5\xee\\\r\x06\xd5\xdd\xf7\xa4+{\x1a\x9f)\xdd\xc0\xdeBcQ\x99\xfd\xb8_\xbe\xeb%\xfco\x16sg\xee#\x18'],
-[b'\xa6\xf9\xf5?C\xf1K\x1cf(A\xaa\xdc\xc6\xf5\xdc\xe1\xdb\xc7\x08\xbfM\x00m\x0e\xc2:\xc0\xb2\xf9\xc5\xbb\x12\xc9\xd0\x13I \xd8\xfc\xfe\xfa\x0b\xaf\x9b\xd5Q\xaaK#\xa4g\x85\x8c\x01\x02\x88\x15Yp\xa6\xfdn\x96\xc4\x93\xec\xd0r\xca|>;6\x9a\xec\x90\x96\xa7\n\xb1y\xcai\x0b\xde\xfe\xfd{h\x03\xb9\x8aj\xba\xae\x98\x12q\xf3\x9d\xb4La\x92\xc4\xd8\xf6!UB\x88\xbf\xf7\x01\x84\xf6O_\x01y^f0\xfa\xfe\xf0r', b'\x9f\nc\x03 \xa0\x7f\x18\xb6\xa4\xda\x82\xd8\xa1k\x8e8e\x0b|\'RZ3\x83R\xd8\xbc\x1bJ"\xa9E+S\xc8\x8e\xa3\xc2\x0f<\xb3D/\xf7\xa8\xde*\xa5<\x85\xa7\x10\xc7\xb1\x9f/\xd0\x0bVxb\xca\xb6\x1dO\xdd0\r\xcb\xf7\xd5\x9d\xbd\xc4Xl\x8d\xcf\xc4\xc5\xc7n6\x04\\\xbaK\x04\x8bg\x9c_\x9d\xe7\xca@\x85\xd2\x95%b\xdd\x99\x8da\x0e\x84\xa7\x15\x95b\x0b\xabR\xf4\x95\'\xdd\x89 t\x90\x8a\xf7\xa9-\xb4'],
-[b'\xa5\x97\x08\x81\xe3G\x19;\xa7\xa0\rcp\x86q\x92,\xea\x89[\xf1\xf7W\xbcl\xe5ro)-l\xec\xa0\xd9w\x07\x9an\xb6\x03\xd3\x93\xae\xca\xd1&\x19\xc0\xe3\x9b\x11\xba\xc6\xacR\xd2.\xb4\x8e\xba}l4\xe1\x0fS\x0b_=\xdd&ZB\xf5]\xdem\xb2\xdb[\xd4\xdd6\x90C\x00\xbe\xa8\xab\t:\x18\x0f\x1b`2vA\x07\x0f\x06\xcf\xdc\x9b\xe8\x12T\x8a C\x83n,\x0b\xa8v\xfd2\xd8\t\xa7\xcf\x8c\xee\xc6\x9fH\xec', b".\xaa\x9e\xa8Ui\xab\xba\x91p_\xb2N\x01\xb1.\x8b\xb1}\xd7B\xc6\xc3\xc9\x07\xb6\xd3D/'T8tYZ\xfcd\xb0v#\xdd\xd4('i\x91z\xf2\xf8i)\xc2\xacu\x97I\x94_\x07.\xe5G\x04}t\\q\x14\xbc\x00'\xfd\x1d\xf6wW#~\x10\xe9r\xd9\xd8&\xa9\x05m\xd7\xf9% \xe9\xea\x00\x9c\x00\xde\xb6\xb6L^$\xb1f\xebp\xba\x12\xe7\x93\xa9\xf8B1/8\xb9\xb2p\xc3\xd4t0\xe3\xaa\xa3)9"],
-[b'\x10\x01\x88\xaa\xb2\x85u\x9c5s\x02\xc0tnO\xa24SW\x80?\x84\x8csXI/\x8c\xefYVF\xef\n\xf3\xe9\x0f\xc9\xb3\xea\x89\xad\xd0]c\xf0\xfd\xe0O\xd5K3-\xb1\x00\x16?\x13\x9dsP3\xc5\xd24\xef\x15\x0bs\x15\x1f\xedw\xc6\xfb5Q\xb3\n.m\xa1\xedF\x9f\xcc\xa3\xec\xe8\xefN\xfa\xc5\xf0+\xa4\x0b\x0f\xdd>E\x9b\xe26\x8c5\xdb\x8e\xa8"\xe2\xec\xca\xf7\xb8G0a\xd8\xf7\xc6\xad3~+\xfe\xc6H', b'Em\x02\xd2\x15}\xe7 7]\x19@)\xb9\xa0u\n9&\xeeK\xad\x1d\xd1\xbe\x96\r\x80\x9b\xcf\xfa8\xe3\x16\xbaH\x1c\xb3\xfd\x16\x8b:\xe4^\x81D\x84,,\xc3\xf3\xcb\xd5\x15\x95=N\x1d\x9a\x87?\x17(`\x97\x0e]]\x999\xdb\x96\xccym+E@\xac\x167C\x07\xdb\xd1\x98\xef\t\xa5\xd4\x81\x9d\x02\xb4\x84\xbd\xecX\x0b\x1e1\xec5\x1c%\x0c|\xb4\x00\x03\x0fF\x9c\xc1\xaf\xd7\x92\xcb\xc2<\xec\xaa\xe0/\ne~v'],
-[b"\x91w[\x1ew#\xc3\x0bN\xefFy,\xdd}\\\xe1\xda\x85\x9a.\x8b\x18\xb7\xf5\xe2=\x81\xac\x91\xe9\x871\x12\xbb\xb0^ \xb7\xca(\x00\x8e\x99\x86\x18\x13\xfb\xa5\x18'[\xe1h\xa0\x80 rmv(\xc8\xdb`\x13-\xb6e'\x1e\xe0\x13N,\xfal\x05\xef'V\xac\xfc\xc7\x00\x981\xa9k\xd1K\xf6_\t\xe3\x85\xeb\x1dvY\x1eQb[@\xb2\xff\xe0\xcaltC\x1c\xbb\xd9%?JS\x82\x88PV\x9bI+\xc6\x87\xa5", b'\xa9^\xd5\xca,|@(\xf9\xf4\xe9Q\xf5\x87\xa5n/\x8e\xa2\xaf\x88\xe1d\xe2B\xef\x14\x8a\xde\x10J=\xf9\xe2\xb7\xf1\xee\xc5\xcd\x02V\xc7\x96\x9b\x13kE\xe8D\xa7$+\x13I\x06\x89\xe6\x1c\xe9Q\xcf\x1d\x06\xddg\x83\xa2\x8d\xcb\xf8\xa6\x8a\x85\xd3\'\xbd\x80.\x84\xbf{\x9d\xcd\xb1\x12\xbbl\xaeG\xc4\x88\xe3\xaf*O"\x11\xfdj\'\xb6\xb5r\xd7\xb4\xd4\xe0,hF\x81\xcc\x9eJ}9\x15\xd8\xaa\x15\xf6n\xeb>)V0\xad'],
-[b"\x17\xb5\xbd\xe3\xe1\x8b\xc0.\xf1\xf8N\xda\xc6\x92L\x94;\x8a\x951R\x98KR\x84\n\xd8\x87\x1e\x1e\xec\xfbK\xfc\xa6\xd8\x86\x7f\xb7\x91\xc80>\x19\x80n\x9a\x0ch\x043\xd44\xf9\xb4\xf7'\xceZ\xfc\x1f\xc9OK\xff\x19\x04h\xc9p\x84\x12X\x9a~\x9e\xc9L\x81\x0b\xcf\xf8\x92l\xee\xce\xed\xfb\xf6:\x96\xc5\x0cz\x19c\xa7\xe4_\xd9\x13\x07\xce\x91S\xb4ko\x08x\xe9\xa5b,/^\xa6\xc9;\xa8\xd9j\xd9G[/\xc3\xbb", b'B\x1b\x91\x8f\xdc\xeb\x93\xeb\x94Qk\x91\xc7E\xe5\x98R\xe4\xbe\x114h\xdd\xf1\x12\x95\x1a&\xa7\x078\xa2\x10Bm\x8b\x0b%7\x18?\xdb\xf4Iv\xb3@\xb9\t\x18m\xd5_\x87\xef\xe8\x93\x92@\xd7\xa1\x19\x1e\xae$\x85rp\xc0\xa5\xa4W\xaa\x89s\x9a\xb0\x8c\xd9Kh\xbbX\x86\x02\xf4\x0f\xbbAN\xa5\xa8i\t!>jXf"8\x81b\xe5p\xe8\x0e\xe0\xb8\\\xb6I\x80\xfa\xbe"\x81\xc9\x90}\x81\xd8\x15\xe0\x90\x0b\xfe\x1f']]
-
-#from mongodb get id and password datas
-def get_accounts():
-    id_list=[]
-    with MongoClient(MONGO_HOST, MONGO_PORT) as client:
-        collection=client.math_questions.account
-        
-        for i in list(collection.find()):
-            phone=utils.rsa_decrypt(i['phone'])
-            pwd=utils.rsa_decrypt(i['password'])
-            id_list.append({'phone':phone,'password':pwd})
-    
-    return id_list
+import json
+#账号文件
+path=r'acounts.json'
 
 if __name__ == "__main__":
-    
-    for i in id_list:
-        
-        id_dict={'phone':i[0],'password':i[1]}
+    f=open(path,'r')
+    accounts_list=json.load(f)
+
+    for i in accounts_list:
+        #加密
+        phone=utils.rsa_encrypt(i[0])
+        pwd=utils.rsa_encrypt(i[0])
+
+        account={'phone':phone,'password':pwd}
         #print(id_dict)
-        mongo_client.insert_one('account',id_dict)
+        mongo_client.insert_one('account',account)
     #print('insert finish')
     
-    id_list=get_accounts()
+    id_list=mongo_client.get_accounts()
     #print(id_list)
